# Optimize deterministic ID creation